refactor: create new component for phone input
