+++ src/components/phone-input/index.ts
@@ -1,0 +1,1 @@
+export {PhoneInput} from './PhoneInput';
